Adding query indexes
Data model is extended with an optional query_index object.
This would not be populated in the normalized SoT models, but
is populated in an extra step. These are intended for query purposes
as well as summary statistics. The closure model from amigo is followed.

Added an extra cli command to be able to index a set.
Also adding example reports